requirements.txt
requests
beautifulsoup4
pandas
lxml
jinja2
pyyaml
